take out old cv interface. just use cv2.
